1. 增加 ApplicationDirectories 以方便创建定制的应用对象 2. theEmptyString, theNullJson 等全局缓存对象，声明为 const 3. Redis 服务默认端口 redis_server_default_port 声明为 const 4. 解决将 Application 前置声明为 struct 的警告 5. 增加全局名字空间下的 _da4qi4_router_xxxx 的字符串后缀操作重载 6. Mount 新 App时，如果发现还没有初始化日志，则自动调用初始化。 7. Application 的查找 for .. in .. 循环的迭代对象，全改为引用（或取会快点？） 8. cookie.cpp 中，解决一处整型精度丢失的警告 9. 解决Response 的 ResplyOK（）的重载冲突 10. Server类增加 AddXXXXsRouter()的三个函数，方便直接使用Server实现多对一路由 11. templates.cpp 中 变量 count 更名为 counts，避免阅读误解

+++ src/cookie.cpp
@@ -8,7 +8,7 @@ namespace da4qi4
 Cookie& Cookie::SetExpires(std::time_t a_time_point)
 {
     std::time_t now(std::time(nullptr));
-    _max_age = a_time_point - now;
+    _max_age = static_cast<int>(a_time_point - now);
     return *this;
 }
